requirements.txt
absl-py==0.8.1
argcomplete==1.10.2
astor==0.8.0
atari-py==0.2.6
attrs==19.3.0
Automat==0.8.0
azure==1.0.3
azure-common==1.1.24
azure-mgmt==0.20.2
azure-mgmt-common==0.20.0
azure-mgmt-compute==0.20.1
azure-mgmt-network==0.20.1
azure-mgmt-nspkg==3.0.2
azure-mgmt-resource==0.20.1
azure-mgmt-storage==0.20.0
azure-nspkg==3.0.2
azure-servicebus==0.20.1
azure-servicemanagement-legacy==0.20.2
azure-storage==0.20.3
backcall==0.1.0
baselines==0.1.5
beautifulsoup4==4.8.2
bleach==3.1.0
boto==2.49.0
box2d-py==2.3.8
cachetools==3.1.1
cairocffi==1.1.0
CairoSVG==2.4.2
cchardet==2.1.6
cffi==1.13.2
chumpy==0.69
Click==7.0
cloudpickle==1.2.2
constantly==15.1.0
crcmod==1.7
cryptography==2.8
cssselect2==0.2.2
cycler==0.10.0
Cython==0.29.14
decorator==4.4.1
defusedxml==0.6.0
dill==0.3.1.1
efficientnet==1.0.0
enlopy==0.1.dev12
eppy==0.5.52
fasteners==0.15
Flask==1.1.2
freetype-py==2.1.0.post1
future==0.18.2
gast==0.2.2
gcs-oauth2-boto-plugin==2.5
gevent==20.4.0
gin-config==0.1.3
glfw==1.10.1
google-apitools==0.5.30
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
google-reauth==0.1.0
greenlet==0.4.15
grpcio==1.25.0
gsutil==4.46
gym==0.12.1
h5py==2.10.0
html5lib==1.0.1
hyperlink==19.0.0
image-classifiers==1.0.0
imageio==2.6.1
imgaug==0.3.0
importlib-metadata==0.23
incremental==17.5.0
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
iso8601==0.1.12
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.0
json5==0.9.2
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
jupyterlab==2.0.1
jupyterlab-server==1.0.7
Keras==2.3.1
Keras-Applications==1.0.8
keras-facenet==0.1a5
Keras-Preprocessing==1.1.0
keras-resnet==0.2.0
keras-retinanet==0.5.1
keras-rl==0.4.2
kiwisolver==1.1.0
lxml==4.5.0
Markdown==3.1.1
matplotlib==3.1.1
mira==0.2
mistune==0.8.4
mock==2.0.0
more-itertools==7.2.0
MouseInfo==0.1.2
mpi4py==3.0.3
munch==2.5.0
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.2
nms==0.1.6
notebook==6.0.2
oauth2client==4.1.3
oauthlib==3.1.0
opencv-contrib-python==4.1.2.30
opendr==0.78
opt-einsum==3.1.0
opyplus==1.0.3
pandas==0.25.3
pandocfilters==1.4.2
pangocffi==0.5.0
parso==0.5.1
pbr==5.4.3
pickleshare==0.7.5
plotly==4.7.0
progressbar2==3.47.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.11.2
psutil==5.6.5
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
PyAudio==0.2.11
PyAutoGUI==0.9.48
pybullet==2.6.5
pycparser==2.19
pydot3k==1.0.17
pyFluidSynth==1.2.5
PyFoam==0.6.11
PyGetWindow==0.0.8
pyglet==1.5.0
Pygments==2.4.2
PyHamcrest==2.0.0
pymodbus==2.3.0
PyMsgBox==1.0.7
pymunk==5.6.0
PyOpenGL==3.1.0
pyOpenSSL==19.1.0
pyparsing==2.4.5
pyperclip==1.7.0
Pyphen==0.9.5
pyphysicssandbox==1.4.1
PyRect==0.1.4
pyrender==0.1.32
pyrsistent==0.15.5
PyScreeze==0.1.26
pyserial==3.4
python-periphery==2.0.1
python-slugify==3.0.6
python-utils==2.3.0
python3-xlib==0.15
PyTweening==1.0.3
pyu2f==0.1.4
PyWavelets==1.1.1
pyzmq==18.1.1
requests-oauthlib==1.3.0
retry-decorator==1.1.0
retrying==1.3.3
rsa==4.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.4.1
seaborn==0.10.0
segmentation-models==1.0.0
Send2Trash==1.5.0
serial==0.0.97
Shapely==1.6.4.post2
smplx==0.1.13
SocksiPy-branch==1.1
sounddevice==0.3.15
SoundFile==0.10.3.post1
soupsieve==2.0
stable-baselines==2.9.0
svgwrite==1.3.1
tabulate==0.8.7
tb-nightly==1.14.0a20190603
tensorboard==2.1.1
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
tf-agents==0.3.0
tf-estimator-nightly==1.14.0.dev2019060501
tinycss2==1.0.2
tinynumpy==1.2.1
tornado==6.0.3
tqdm==4.38.0
traitlets==4.3.3
trimesh==3.4.4
Twisted==19.10.0
Unidecode==1.1.1
validators==0.14.1
wcwidth==0.1.7
WeasyPrint==51
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0
zmq==0.0.0
zope.interface==4.7.1
